pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "enhanced-cdar"
version = "0.1.0"
description = "Enhanced Conditional Drawdown at Risk modeling toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
dependencies = [
  "numpy>=1.24,<3",
  "pandas>=2.0,<3",
  "scipy>=1.10,<2",
  "cvxpy>=1.4,<2",
  "plotly>=5.18,<6",
  "matplotlib>=3.7,<4",
  "pydantic>=2.5,<3",
  "PyYAML>=6.0,<7",
  "typer>=0.12,<1",
  "yfinance>=0.2.40,<1",
  "joblib>=1.4,<2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
  "pytest-cov>=5.0,<6",
  "hypothesis>=6.0,<7",
  "ruff>=0.6,<1",
  "black>=24.0,<25",
  "isort>=5.13,<6",
  "mypy>=1.10,<2",
]

[project.scripts]
enhanced-cdar = "enhanced_cdar.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q --cov=enhanced_cdar --cov-report=term-missing"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_return_any = true
disallow_untyped_defs = false
ignore_missing_imports = true
